pyproject.toml
# Python package dependencies
[project]
name = "xic-rzerocot"
version = "0.1.0"
description = "Executable-Invariant Certificate Reflective Zero-shot CoT"
requires-python = ">=3.9"

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "hydra-core>=1.3.0",
    "wandb>=0.15.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "omegaconf>=2.3.0",
    "accelerate>=0.20.0",
]